poll-like commands now don't delete original message when bot is in test mode

+++ commands/ask.js
@@ -40,7 +40,7 @@ class AskCommand extends Command {
 			await sent.react(config.yes_react);
 			sent.react(config.no_react);
 		}
-		if (message.channel.type != "dm") {
+		if (message.channel.type != "dm" && !this.client.testMode) {
 			return await message.delete();
 		}
 	}
